Fix compare file path (#20)
* Fix Issue #18
Change comparison so that if files match we return true.

* Created Compare-File function

Renamed Compare-FilePath to Compare-File. Recreated Compare-FilePath to return negated Compare-File with a Write-Warning that Compare-FilePath is deprecated. This seems to be the best option to not break code that relies on Compare-FilePath returning the incorrect data.

Don't forget to export the function from the module.